+++ ShoutOut.Core/Entities/ShoutImage.cs
@@ -1,4 +1,4 @@
-﻿namespace ShoutOut.Core
+﻿namespace ShoutOut.Core.Entities
 {
     public class ShoutImage : Entity
     {
